VERSION casi final del proyecto

+++ Project_1/Models/Propiedad_Del_Propietario/Propiedad_del_Propietario_Conexion.cs
@@ -106,7 +106,7 @@ namespace Project_1.Models
                             propietarios.Add(new Propietario()
                             {
                                 nombre = reader.GetString(0),
-                                valorDocId = reader.GetInt32(1),
+                                valorDocId = (int) reader.GetInt64(1),
                                 nombreDocId = reader.GetString(2)
 
                             });
